refactor: mock.ts, seed.ts migration
